Update Quarto report for 2025 model (#48)
This PR updates the Quarto report to default to the 2025 model. In the
process, it also streamlines the Quarto params to tie final models to
years so that the user only has to set the `year` for the report they
would like to generate and the doc will infer the correct model run ID
based on the year.