Fix info box rendering in README (#63)
While auditing the code in this repo for sensitive information, I
noticed that the info box in the README is using the wrong shortcode,
causing it to render incorrectly. This PR fixes that problem. You can
verify the fix by switching to the "Preview" mode in the diff and
confirming that the info box looks correct.