Stop booleans from getting cast to strings, resolve map render bug (#67)
When we added `_format_dict_numbers()` we weren't excluding
`"has_subject_pin_sale"` and `"is_subject_pin_sale"`. This cast them to
string type which screwed up the conditional flow in the html doc, which
messed up the map rendering tooltips. This now works, and will be
re-worked to be simpler in https://github.com/ccao-data/pinval/issues/62