Refactor workflow to accept `year` input and invalidate cache on deploy (#74)
## Background

This PR undertakes a major refactor of the `generate-pinval` workflow to
support three main goals:

1. Invalidate the CloudFront cache on deploy
(https://github.com/ccao-data/pinval/issues/68) so changes are visible
immediately in the live site
2. Accept an optional `year` argument that will select the final model
run for the given assessment year, in contrast to `run_id`, which
requires the caller to know the name of the run ID they want to use
3. Clean up and reorganize the workflow so that all parameter validation
and processing takes place in a preliminary `setup` stage that all
following stages depend on

Goal 1 started out as the main goal of this PR, but I quickly noticed a
pre-existing issue with the workflow: We need to know the assessment
year in order to save files to S3 and invalidate the cache based on S3
paths, but we had been deriving the assessment year from the model run
ID, which is not technically correct. Attempting to extract assessment
years for model run IDs led me to realize that the workflow would be a
lot clearer to read and easier to use if we tackled goals 2 and 3. The
result is this PR, which is a much bigger diff than I had expected, but
in the end I think it will make for a workflow that is easier to update
and use.

## Testing

Since this PR makes major changes to the pipeline, I tested every
possible combination of inputs to make sure they all behave as expected:

| args | behavior | logs |
| ---- | ---------- | ---- |
| `year=2025` | updates all PINs in north tri townships** |
[logs](https://github.com/ccao-data/pinval/actions/runs/15885569997/job/44796506231)
|
| `run_id=2025-02-11-charming-eric` | updates all PINs in north tri
townships** |
[logs](https://github.com/ccao-data/pinval/actions/runs/15885626044/job/44796697686)
|
| `year=2025`<br>`pins="01014030300000 01014090050000"` | updates both
PINs |
[logs](https://github.com/ccao-data/pinval/actions/runs/15883812682/job/44790705768)
|
| `run_id=2025-02-11-charming-eric`<br>`pins="01014030300000
01014090050000"` | updates both PINs |
[logs](https://github.com/ccao-data/pinval/actions/runs/15885348940/job/44795786491)
|
| `year=2025`<br>`run_id=2025-02-11-charming-eric` | raises an error
(mutually exclusive args) |
[logs](https://github.com/ccao-data/pinval/actions/runs/15885481502/job/44796215907)
|
| `year=1234` | raises an error (invalid year) |
[logs](https://github.com/ccao-data/pinval/actions/runs/15885494231/job/44796257111)
|
| `run_id=2020-01-01-fake-run` | raises an error (invalid run ID) |
[logs](https://github.com/ccao-data/pinval/actions/runs/15885546210/job/44796432078)
|
| `pins=1234567890` | raises an error (invalid PIN) |
[logs](https://github.com/ccao-data/pinval/actions/runs/15886246222/job/44798830083)
|
| No arguments | raises an error (run ID or year is required) |
[logs](https://github.com/ccao-data/pinval/actions/runs/15886412199/job/44799368760)
|

**_Note: I canceled this job after confirming that it correctly parsed
all arguments, kicked off the expected township job matrix, and issued
the correct SQL queries; my thinking is that it wasn't worth actually
updating all staging reports to test the final steps of this job since
the final steps are very similar to the other test runs_