Add error pages for ineligible parcels and bad requests (#79)
## Background

This PR adds support for two types of error pages to the Hugo site:

- **Ineligible parcels**: We will generate these reports when the PIN
and year are valid, but the PIN isn't eligible for a PINVAL report
- In this case, the report will display a user-friendly explanation of
why the PIN isn't eligible in a given year, along with a contact form
that the user can use to contact us if they believe the parcel _should_
be eligible in that year
- **Bad requests**: We will generate one page for this generic error and
set it as [the default page for AWS to redirect
to](https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/creating-custom-error-pages.html)
in case the user requests a URL that does not match a report file
- In this case, the user won't get any explanation for why the report
doesn't exist, because we won't have any information about what they did
wrong; I think this is fine because the page is fundamentally a fallback
to provide a friendlier experience in case of malformed URLs, and a user
who is using the web form to access a report should not face a malformed
URL

Some other small changes that I folded into this PR for the sake of
convenience:

- **Adding a township workflow input to only deploy reports for one
specific township**. We had been discussing this feature anyway, and I
realized I needed it to help debug memory issues in Lake township, so I
went ahead and added it because it was easy.
- **Adding a boolean workflow input to control whether we deploy static
assets like favicons and the generic error page**. Previously, we didn't
have a way of deploying these assets automatically.
- **Deleting old example reports that we used for intermediate
testing**.
- **Adding a `robots.txt` file to block all web crawlers**. We don't
want search engines indexing these reports, since we want to encourage
users to go through the flow of using the web form on
cookcountyassessor.com.
- **Slightly more memory efficient iteration in the `build_front_matter`
loop in `generate_pinval.py`.** I needed to do this because we were just
barely running out of memory when generating reports for Lake township.

Connects https://github.com/ccao-data/pinval/issues/51.

## Testing

Workflow runs that succeeded:

- [Full 2025 run on all PINs in the
County](https://github.com/ccao-data/pinval/actions/runs/16054982811)
- [Township filter for just
Barrington](https://github.com/ccao-data/pinval/actions/runs/16056095114)

Here are some PINs that I've tested:

| pin	| parcel_class | reason_report_ineligible | on staging? |
| --- | ---- | ---- | --- |
| 17101090241016 | 299	| `condo` |  ✅ |
| 13352300250000 | 212 | `non_tri` |  ✅  |
| 25271150140000 | EX | `non_regression_class` |  ✅|
| 04061010200000 | 203 | `missing_card` |  ✅ |
| 01014030300000 | 208 |  |  ✅ |
| 00000000000000 |        | 404 | ✅ |