Fix average comps sale price bug (#83)
When calculating the average sale price and sale price per square foot
for the comparables, we were only grouping by `card`. When running the
python script for more than one PIN, this caused the group by operation
to grab the mean for all of the PINs with card "1" rather that just the
comps for a single PIN, for example. In this PR we add `pin` as a
grouping variable to fix this.