Refactor to use new simplified `pinval.vw_assessment_card` data model (#85)
This PR updates our Hugo report generation code to handle the changes to
the `pinval.vw_assessment_card` data model that we're making in
https://github.com/ccao-data/data-architecture/pull/856.

In addition, we fold in one small change to improve the experience of
testing changes to the data model: Factoring out constants for
`PINVAL_ASSESSMENT_CARD_TABLE` and `PINVAL_COMP_TABLE`. By centralizing
the references to these two tables, we can more easily change them when
we need to test a corresponding change in `data-architecture`.

See PIN `06171140020000` on the staging site for an example of a
deployed report.

Connects https://github.com/ccao-data/data-architecture/issues/855.