Add "Back to search" button option (#95)
This PR adds a "back to search" button which functions like a back
button, linking back to the home page of the model report website where
the user can search for a year and pin. It adds an `environment` flag to
the script, which allows us to specify if we want to switch back to the
dev or prod version of the website.

Closes https://github.com/ccao-data/pinval/issues/84

---------

Co-authored-by: Jean Cochrane <jeancochrane@users.noreply.github.com>