Finalize table controls and improve small+recent multicard UX (#96)
This PR makes a few changes to finalize characteristic table controls
and improve the UX for small recent (>2025) multicards.

Table controls now look like this:
- Search on the top left
- "Export" button on the top right, which drops down to three options
("Copy", "Excel", and "Print View")
- Sticky footer with two buttons: "Next section", which jumps to after
the table, and "Previous section", which jumps to before the table

Plus, the table now uses the [Responsive
extension](https://datatables.net/extensions/responsive/) so that extra
columns get folded into a row dropdown on small viewports.
 
For small recent (>2025) multicards, the card tabset now flips between
chars for the cards, but the map and the characteristic table are no
longer part of the tabset. This is intended to better communicate that
all cards in small recent multicards share the same comps and
characteristics for valuation purposes.

We also fold a few small changes into this PR for convenience:

- Adds a "Back to top" button to the bottom of the page
- Switches the info icon in the "Read more" banner to use a Bootstrap
icon, like the rest of the icons in the report
- Updates DataTables CSS/JS dependencies to make sure we're correctly
loading the required Bootstrap-enabled versions
- Fix some minor indentation problems
- Use the text "five" instead of the numeric "5" in all paragraph text
referring to "top five sales"
- Alert styles for supplemental info that we want to call the user's
attention to (multicard explainer and desk review warning)

Closes #97, #43

---------

Co-authored-by: Jean Cochrane <jean@jeancochrane.com>