Add support for 2024 comps to Hugo report and fix small multicard bugs (#104)
The main goal of this PR, in collaboration with its companion
https://github.com/ccao-data/data-architecture/pull/864, is to enable
generating comps for the 2024 assessment year.

We also roll in a few smaller changes that make sense to do at the same
time:

- Switches to using the comps model run ID as the basis for the
`generate-pinval` script and workflow, to match data model changes made
in https://github.com/ccao-data/data-architecture/pull/864 that are
necessary to support multiple final models per year (as in 2024)
- Switches to using the `is_frankencard` card attribute to determine
which card is the frankencard for small+recent multicards
- Our previous approach (using the first card) sometimes chose the
incorrect card when cards weren't ordered in descending order of square
footage, so this change fixes that bug
- Resolves two small bugs related to multicard display as described in
#103:
- Say "home" instead of "card" when describing the combined initial
model esimate for small+recent multicards
- Modify the $/sqft calculation for small+recent multicards to use the
combined building square footage

Finally, we make a few small tweaks to clean up the code, which are not
relevant enough to list here; I'll document them in comments below.

Example PINs for testing:

* 2024
    * One card: 10314090100000
    * Two cards: 13363050390000
    * Three cards: 16251250220000
    * Four cards: 25223030460000
    * Five cards: 13234080750000
* 2025
    * One card: 05271130010000
    * Two cards: 01011240390000
    * Three cards: 09092000170000
    * Four cards: 06041000130000

Closes https://github.com/ccao-data/pinval/issues/75, #81,
https://github.com/ccao-data/pinval/issues/103