Add a "Score" column to the characteristics table to display feature importance (#113)
This PR updates the report to add a new "Score" column to the
characteristics table that shows feature importance on a scale of 1-4,
with stars illustrating the scores.

This new feature is powered by two primary changes:

- Updates the `generate_pinval.py` script to use the SHAPs that we added
to the `pinval.vw_assessment_card` view in
https://github.com/ccao-data/data-architecture/pull/875 to compute
relative weights for each feature, as well as quartiles based on the
weights
- Updates the `report.html` file to display these quartiles in the
characteristics table using stars, with weights included to allow the
user to sort by true feature importance.

See the following testing PINs on the staging site:

- 2024
    - Single card: 10314090100000
    - Two card: 13363050390000
- 2025
    - Single card: 05271130010000
    - Two card: 01011240390000

Closes #86.