Update social card with new report name (#124)
This PR updates the social cards we added in #115 to reflect the new
name for the reports.

To test, enter the staging site link for PIN 01011000040000 into a Teams
chat with yourself, which will render the text that we changed here as
the title. You can also test the PIN on https://metatags.io/ though you
should note that most social platforms do not actually display the title
that we are changing here, hence that website is not particularly useful
for testing in this case.