Small tweaks to link refs and basemap styles (#125)
This PR folds in a couple fixes for small bugs that we noticed while
testing prod reports:

- Changes the Carto basemap from Positron to Voyager for a more colorful
map that is closer to users' expectations of a typical web map
- Adds "Back to Search" buttons to error pages
- Changes link refs for the CCAO site to point to the new `.gov` domain

## Testing

I deployed all of the PINs that are listed in #116 to staging to confirm
that links and maps appear correctly in all possible versions of the
report. To test the error page, visit the root domain of the staging
site.